requirements.txt
kafka-python
pyspark==3.4.1
pandas
plotly
dash
